Fix credits
Because the value begins with an opening square bracket `[`, the YAML parser wants to read it as an array, and then chokes when it hits the first opening parenthesis.  One way to deal with this is to quote the hole string.  Another is to make it a multi-line string, as is done here.  Either way the effect is to force the parser to read it as a string value (which we will subsequently parse as markdown).